Support shapefile->postgis conversion and charset parameter

+++ src/geoserver/catalog.py
@@ -1,4 +1,6 @@
 from datetime import datetime, timedelta
+import logging
+from django.core.files.uploadedfile import InMemoryUploadedFile
 from geoserver.layer import Layer
 from geoserver.store import DataStore, CoverageStore
 from geoserver.style import Style
@@ -6,10 +8,9 @@ from geoserver.support import prepare_upload_bundle
 from geoserver.layergroup import LayerGroup
 from geoserver.workspace import Workspace
 from os import unlink
-import httplib2 
+import httplib2
 from xml.etree.ElementTree import XML
 from urlparse import urlparse
-import logging
 
 class UploadError(Exception):
     pass
@@ -18,7 +19,7 @@ class ConflictingDataError(Exception):
     pass
 
 class AmbiguousRequestError(Exception):
-    pass 
+    pass
 
 class FailedRequestError(Exception):
     pass
@@ -33,7 +34,7 @@ class Catalog(object):
   - Layers, which combine styles with resources to create a visible map layer
   - LayerGroups, which alias one or more layers for convenience
   - Workspaces, which provide logical grouping of Stores
-  - Maps, which provide a set of OWS services with a subset of the server's 
+  - Maps, which provide a set of OWS services with a subset of the server's
     Layers
   - Namespaces, which provide unique identifiers for resources
   """
@@ -49,12 +50,12 @@ class Catalog(object):
     netloc = urlparse(url).netloc
     self.http.authorizations.append(
         httplib2.BasicAuthentication(
-            (username, password), 
+            (username, password),
             netloc,
             url,
             {},
             None,
-            None, 
+            None,
             self.http
             ))
     self._cache = dict()
@@ -67,8 +68,8 @@ class Catalog(object):
 
   def delete(self, object, purge=False):
     """
-    send a delete request 
-    XXX [more here]   
+    send a delete request
+    XXX [more here]
     """
     url = object.get_url(self.service_url)
 
@@ -78,12 +79,12 @@ class Catalog(object):
     headers = {
       "Content-type": "application/xml",
       "Accept": "application/xml"
-    } 
+    }
     response = self.http.request(url, "DELETE", headers=headers)
     self._cache.clear()
     return response
 
-  
+
   def get_xml(self, url):
     cached_response = self._cache.get(url)
 
@@ -96,6 +97,7 @@ class Catalog(object):
         response, content = self.http.request(url)
         if response.status == 200:
             self._cache[url] = (datetime.now(), content)
+            logging.info(url)
             return XML(content)
         else:
             raise FailedRequestError("Tried to make a GET request to %s but got a %d status code: \n%s" % (url, response.status, content))
@@ -118,6 +120,7 @@ class Catalog(object):
     return response
 
   def get_store(self, name, workspace=None):
+      logging.debug("Search for store [%s]", name )
       stores = [s for s in self.get_stores(workspace) if s.name == name]
       if workspace is None:
           store = None
@@ -138,6 +141,7 @@ class Catalog(object):
               raise FailedRequestError("No store found named: " + name)
           return store
       else: # workspace is not None
+          logging.debug("datastore url is [%s]", workspace.datastore_url )
           ds_list = self.get_xml(workspace.datastore_url)
           cs_list = self.get_xml(workspace.coveragestore_url)
           datastores = [n for n in ds_list.findall("dataStore") if n.find("name").text == name]
@@ -167,7 +171,8 @@ class Catalog(object):
               stores.extend(a)
           return stores
 
-  def create_pg_feature(self, storename, name, data, workspace=None, overwrite=False):
+
+  def create_pg_feature(self, storename, name, data, workspace=None, overwrite=False, charset=None):
 
     if not overwrite:
         try:
@@ -187,7 +192,10 @@ class Catalog(object):
     store = self.get_store(storename, workspace)
 
 
-    ds_url = "%s/workspaces/%s/datastores/%s/file.shp" % (self.service_url, workspace.name, storename)
+    if charset:
+        ds_url = "%s/workspaces/%s/datastores/%s/file.shp?charset=%s" % (self.service_url, workspace.name, storename, charset)
+    else:
+        ds_url = "%s/workspaces/%s/datastores/%s/file.shp" % (self.service_url, workspace.name, storename)
     # PUT /workspaces/<ws>/datastores/<ds>/file.shp
     headers = {
       "Content-type": "application/zip",
@@ -204,14 +212,15 @@ class Catalog(object):
       if headers.status != 201:
           raise UploadError(response)
     finally:
+      if zip:
         unlink(zip)
 
-  def create_featurestore(self, name, data, workspace=None, overwrite=False):
+  def create_featurestore(self, name, data, workspace=None, overwrite=False, charset=None):
     if not overwrite:
         try:
             store = self.get_store(name, workspace)
             msg = "There is already a store named " + name
-            if workspace: 
+            if workspace:
                 msg += " in " + str(workspace)
             raise ConflictingDataError(msg)
         except FailedRequestError, e:
@@ -220,7 +229,11 @@ class Catalog(object):
 
     if workspace is None:
       workspace = self.get_default_workspace()
-    ds_url = "%s/workspaces/%s/datastores/%s/file.shp" % (self.service_url, workspace.name, name)
+    if charset:
+        ds_url = "%s/workspaces/%s/datastores/%s/file.shp?charset=%s" % (self.service_url, workspace.name, name, charset)
+    else:
+        ds_url = "%s/workspaces/%s/datastores/%s/file.shp" % (self.service_url, workspace.name, name)
+
     # PUT /workspaces/<ws>/datastores/<ds>/file.shp
     headers = {
       "Content-type": "application/zip",
@@ -241,7 +254,7 @@ class Catalog(object):
         try:
             store = self.get_store(name, workspace)
             msg = "There is already a store named " + name
-            if workspace: 
+            if workspace:
                 msg += " in " + str(workspace)
             raise ConflictingDataError(msg)
         except FailedRequestError, e:
@@ -340,7 +353,7 @@ class Catalog(object):
     raise NotImplementedError()
 
   def get_layergroup(self, id=None, name=None):
-    group = self.get_xml("%s/layergroups/%s.xml" % (self.service_url, name))    
+    group = self.get_xml("%s/layergroups/%s.xml" % (self.service_url, name))
     return LayerGroup(self, group.find("name").text)
 
   def get_layergroups(self):
@@ -380,7 +393,7 @@ class Catalog(object):
 
     self._cache.clear()
     if headers.status < 200 or headers.status > 299: raise UploadError(response)
-  
+
   def get_namespace(self, id=None, prefix=None, uri=None):
     raise NotImplementedError()
